requirements.txt
flask==3.1.0
boto3==1.38.13
werkzeug==3.1.3
python-dotenv==1.1.0
pillow==11.2.1
pillow-heif==0.22.0
gunicorn==23.0.0